% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_colour_seqrise}
\alias{scale_colour_seqrise}
\alias{scale_fill_seqrise}
\alias{scale_color_seqrise}
\title{Seqrise color scales.}
\usage{
scale_colour_seqrise(palette = "seqrise9", ...)

scale_fill_seqrise(palette = "seqrise9", ...)

scale_color_seqrise(palette = "seqrise9", ...)
}
\arguments{
\item{palette}{Palette name.}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Seqrise color scales.
}
\seealso{
\code{\link{seqrise_color_pal}} for references.

Other colour seqrise: \code{\link{scale_colour_gradient2_seqrise}},
  \code{\link{scale_colour_gradient_seqrise}},
  \code{\link{seqrise_color_pal}},
  \code{\link{seqrise_div_gradient_pal}},
  \code{\link{seqrise_seq_gradient_pal}}
}
