% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{seqrise_color_pal}
\alias{seqrise_color_pal}
\title{Color Palettes based on Seqrise (discrete)}
\usage{
seqrise_color_pal(palette = "seqrise9")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color Palettes based on Seqrise (discrete)
}
\seealso{
Other colour seqrise: \code{\link{scale_colour_gradient2_seqrise}},
  \code{\link{scale_colour_gradient_seqrise}},
  \code{\link{scale_colour_seqrise}},
  \code{\link{seqrise_div_gradient_pal}},
  \code{\link{seqrise_seq_gradient_pal}}
}
