% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{seqrise_div_gradient_pal}
\alias{seqrise_div_gradient_pal}
\title{Seqrise diverging colour gradient palettes (continuous)}
\usage{
seqrise_div_gradient_pal(palette = "Blue-White-Red", space = "rgb")
}
\arguments{
\item{palette}{Palette name. See \code{seqriseR_data$seqrise$divergent}.}

\item{space}{Colour space in which to calculate gradient.}
}
\description{
Seqrise diverging colour gradient palettes (continuous)
}
\seealso{
Other colour seqrise: \code{\link{scale_colour_gradient2_seqrise}},
  \code{\link{scale_colour_gradient_seqrise}},
  \code{\link{scale_colour_seqrise}},
  \code{\link{seqrise_color_pal}},
  \code{\link{seqrise_seq_gradient_pal}}
}
