% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqriseR-data.R
\docType{data}
\name{seqriseR_data}
\alias{seqriseR_data}
\title{Palette data for the seqriseR package}
\format{A \code{list}.}
\usage{
seqriseR_data
}
\description{
Data used by the palettes in the seqriseR package.
}
\keyword{datasets}
