% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-base.R
\name{theme_base}
\alias{theme_base}
\title{The default font size is set to 14
The default font family is set to Helvetica}
\usage{
theme_base(base_size = 14, base_family = "Helvetica")
}
\arguments{
\item{base_size}{\code{integer} basic font size}

\item{base_family}{\code{string} basic font family}
}
\description{
The default font size is set to 14
The default font family is set to Helvetica
}
